Change "Interrupted" to C-q, and avoid interrupting the editor.
Because C-c is muscle-memory bound to "copy" (despite Y, which sometimes
doesn't work (eg. while in insert mode), although
http://phaseportrait.blogspot.com.au/2008/12/make-vimperators-yank-behave-like-edit.html
helps a lot), using it alongside C-i for the external textarea editor is
diabolically bad, and has caused me to lose many edits (to the point that my
editor command is now a script which backs up the file being edited on the way
in and the way out of the actual editor).

The bottom line of all this is that I hate the behaviour of C-c in vimperator,
and would prefer C-c stay as "Copy".

So, this patch adjusts that behaviour, and also uses C-q instead of C-c to
interrupt things (but not the internal editor).